requirement.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.4
SQLAlchemy==2.0.35
pymysql==1.1.2
cryptography==46.0.3
mysql-connector-python==9.5.0